naturesrpg/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-filter==2.4.0
djangorestframework==3.12.4
httplib2==0.19.1
idna==2.10
jwcrypto==0.8
keyring==21.4.0
oauth2==1.9.0.post1
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
pyinaturalist==0.12.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-forge==18.6.0
pytz==2021.1
requests==2.25.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4